+++ src/config/config.test.js
@@ -33,7 +33,7 @@ describe('Config ', () => {
           AUTH_TOKENS: '/auth/token',
           AUTH_REFRESH: '/auth/refresh',
           AVAILABLE_BALANCE: '/available_balance',
-          COMMITTED_AMOUNTS: '/committed_amounts',
+          COMMITTED_AMOUNT: '/committed_amount',
           USER_DATA: '/user_data'
         },
         relay: {