+++ src/the-ocean-wallet.js
@@ -11,39 +11,39 @@ export default class Wallet {
   /**
    * To get how many tokens in an account
    * @param params
-   * @param  {String}  params.etherAddress The owner address
+   * @param  {String}  params.walletAddress The owner address
    * @param  {String}  params.tokenAddress The string address of the token
    * @return Promise<BigNumber> The bignumber with your balance
    */
-  async getTokenBalance ({etherAddress, tokenAddress}) {
-    const address = etherAddress || this.web3.eth.defaultAccount
+  async getTokenBalance ({walletAddress, tokenAddress}) {
+    const address = walletAddress || this.web3.eth.defaultAccount
     return this.zeroEx.token.getBalanceAsync(tokenAddress, address)
   }
 
   /**
    * To get how many tokens have you allowed to the relayer
    * @param params
-   * @param  {String}  params.etherAddress The owner address that allowed the relayer
+   * @param  {String}  params.walletAddress The owner address that allowed the relayer
    * @param  {String}  params.tokenAddress The string address of the token
    * @return Promise<BigNumber> The bignumber with your allowance
    */
-  async getTokenAllowance ({etherAddress, tokenAddress}) {
-    const address = etherAddress || this.web3.eth.defaultAccount
+  async getTokenAllowance ({walletAddress, tokenAddress}) {
+    const address = walletAddress || this.web3.eth.defaultAccount
     return this.zeroEx.token.getProxyAllowanceAsync(tokenAddress, address)
   }
 
   /**
    * To give an allowance to the relayer for later use
    * @param params
-   * @param  {String}  params.etherAddress The address of your ETH account
+   * @param  {String}  params.walletAddress The address of your ETH account
    * @param  {String}  params.tokenAddress The string address of the token
    * @param  {BigNumber}  params.amountInWei How many tokens you want to allow.
    * @callback {submittedCallback} [params.onSubmit] The callback to be called after transaction send to the network
    * @return Promise<Boolean> Operation status
    */
-  async setTokenAllowance ({etherAddress, tokenAddress, amountInWei, onSubmit}) {
+  async setTokenAllowance ({walletAddress, tokenAddress, amountInWei, onSubmit}) {
     const bigAmount = new BigNumber(this.web3.fromWei(amountInWei))
-    const address = etherAddress || this.web3.eth.defaultAccount
+    const address = walletAddress || this.web3.eth.defaultAccount
     const txHash = await this.zeroEx.token.setProxyAllowanceAsync(tokenAddress, address, bigAmount)
 
     if (onSubmit) onSubmit()
@@ -59,13 +59,13 @@ export default class Wallet {
   /**
    * To allow a gigantestic amount of tokens for the relayer, considered unlimited
    * @param params
-   * @param  {String}  params.etherAddress The address of your ETH account
+   * @param  {String}  params.walletAddress The address of your ETH account
    * @param  {String}  params.tokenAddress The string address of the token
    * @callback {submittedCallback} [params.onSubmit] The callback to be called after transaction send to the network
    * @return Promise<Boolean> Operation status
    */
-  async setTokenAllowanceUnlimited ({etherAddress, tokenAddress, onSubmit}) {
-    const ownerAddress = etherAddress || this.web3.eth.defaultAccount
+  async setTokenAllowanceUnlimited ({walletAddress, tokenAddress, onSubmit}) {
+    const ownerAddress = walletAddress || this.web3.eth.defaultAccount
     await this.zeroEx.token.setUnlimitedProxyAllowanceAsync(tokenAddress, ownerAddress)
 
     if (onSubmit) onSubmit()
@@ -82,10 +82,10 @@ export default class Wallet {
    * @return Promise<Boolean> Operation status
    */
   async wrapEth ({amountInWei, address, onSubmit}) {
-    const etherAddress = await this.zeroEx.tokenRegistry.getTokenAddressBySymbolIfExistsAsync('WETH')
+    const walletAddress = await this.zeroEx.tokenRegistry.getTokenAddressBySymbolIfExistsAsync('WETH')
     const bigAmount = new BigNumber(amountInWei)
     const depositor = address || this.web3.eth.defaultAccount
-    const txHash = await this.zeroEx.etherToken.depositAsync(etherAddress, bigAmount, depositor)
+    const txHash = await this.zeroEx.etherToken.depositAsync(walletAddress, bigAmount, depositor)
 
     if (onSubmit) onSubmit()