+++ src/ws/streams/user-data-stream.js
@@ -1,7 +1,7 @@
 import BaseUserStream from './base-user-stream'
 import { CHANNEL } from './constants'
 
-const debug = require('debug')('the-ocean-x:stream:user_data')
+const debug = require('debug')('the-ocean:stream:user_data')
 
 export default class UserDataStream extends BaseUserStream {
   constructor (io) {