+++ test/the-ocean-trade.test.js
@@ -1,6 +1,6 @@
 /* global expect, describe, it, beforeEach  */
 
-import { Trade } from './../src/the-ocean-x-trade'
+import { Trade } from '../src/the-ocean-trade'
 
 describe('Trade ', () => {
   let trade
@@ -14,5 +14,7 @@ describe('Trade ', () => {
     expect(trade.newLimitOrder).toBeDefined()
     expect(trade.cancelOrder).toBeDefined()
     expect(trade.userHistory).toBeDefined()
+    expect(trade.tokenAvailableBalance).toBeDefined()
+    expect(trade.tokenCommittedAmount).toBeDefined()
   })
 })