+++ test/the-ocean.test.js
@@ -1,7 +1,7 @@
 /* global jest, expect, describe, it, beforeAll  */
 
 import { setConfig } from '../src/config/config'
-import createTheOceanX from '../src/the-ocean-x'
+import createTheOcean from '../src/the-ocean'
 import Web3 from 'web3'
 
 describe('MarketData ', () => {
@@ -16,7 +16,7 @@ describe('MarketData ', () => {
 
   it('warns about initialization without web3 provider and exposes only marketData & stream methods', async () => {
     console.warn = jest.fn(warn => {})
-    const api = await createTheOceanX({
+    const api = await createTheOcean({
       api: {
         baseURL: process.env.API_URL
       }
@@ -34,7 +34,7 @@ describe('MarketData ', () => {
   it('warns about initialization without authentication and hides trade methods', async () => {
     console.warn = jest.fn(warn => {})
     let web3Provider = new Web3.providers.HttpProvider('https://kovan.infura.io/XGsH1k50zyZmp6J5CwUz')
-    const api = await createTheOceanX({
+    const api = await createTheOcean({
       web3Provider: web3Provider,
       api: {
         baseURL: process.env.API_URL
@@ -53,7 +53,7 @@ describe('MarketData ', () => {
   it('exposes trade api if API key and secret are provided', async () => {
     console.warn = jest.fn(warn => {})
     let web3Provider = new Web3.providers.HttpProvider('https://kovan.infura.io/XGsH1k50zyZmp6J5CwUz')
-    const api = await createTheOceanX({
+    const api = await createTheOcean({
       web3Provider: web3Provider,
       api: {
         baseURL: process.env.API_URL,
@@ -74,7 +74,7 @@ describe('MarketData ', () => {
   it('exposes trade api if dashboard user tokens are provided', async () => {
     console.warn = jest.fn(warn => {})
     let web3Provider = new Web3.providers.HttpProvider('https://kovan.infura.io/XGsH1k50zyZmp6J5CwUz')
-    const api = await createTheOceanX({
+    const api = await createTheOcean({
       web3Provider: web3Provider,
       api: {
         baseURL: process.env.API_URL